making veterinary reports and animal view

+++ app/Http/Controllers/FeedingReportController.php
@@ -6,16 +6,33 @@ use App\Http\Requests\StoreFeedingReportRequest;
 use App\Models\Animal;
 use App\Models\FeedingReport;
 use App\Models\User;
+use Illuminate\Http\Request;
 
 class FeedingReportController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $feedingReports = FeedingReport::all();
-        return view('feeding-reports.index', compact('feedingReports'));
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if ($user->isVeterinary()) {
+            $query = FeedingReport::latest();
+            $animalId = $request->input('for', null);
+            if ($animalId) {
+                $animal = Animal::findOrFail($animalId);
+                $query->where('animal_id', $animal->id);
+            } else {
+                $animal = null;
+            }
+            $feedingReports = $query->get();
+            return view('vet.feeding-reports.index', compact(['feedingReports', 'animal']));
+        } elseif ($user->isAdmin()) {
+            $feedingReports = FeedingReport::all();
+            return view('feeding-reports.index', compact('feedingReports'));
+        }
+        abort(403);
     }
 
     /**
@@ -63,7 +80,14 @@ class FeedingReportController extends Controller
      */
     public function show(FeedingReport $feedingReport)
     {
-        return view('feeding-reports.show', compact('feedingReport'));
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if ($user->isVeterinary()) {
+            return view('vet.feeding-reports.show', compact('feedingReport'));
+        } elseif ($user->isAdmin()) {
+            return view('feeding-reports.show', compact('feedingReport'));
+        }
+        abort(403);
     }
 
     /**
@@ -71,6 +95,11 @@ class FeedingReportController extends Controller
      */
     public function edit(FeedingReport $feedingReport)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin()) {
+            abort(403);
+        }
         $employees = User::employee()->get();
         $animals = Animal::all();
         return view('feeding-reports.edit', compact('feedingReport', 'employees', 'animals'));
@@ -81,6 +110,12 @@ class FeedingReportController extends Controller
      */
     public function update(StoreFeedingReportRequest $request, FeedingReport $feedingReport)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin()) {
+            abort(403);
+        }
+
         $feedingReport->update($request->validated());
 
         return redirect()->route('feeding-reports.index')->with('success', 'Feeding report updated successfully.');
@@ -91,6 +126,12 @@ class FeedingReportController extends Controller
      */
     public function destroy(FeedingReport $feedingReport)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin()) {
+            abort(403);
+        }
+
         $feedingReport->delete();
 
         return redirect()->route('feeding-reports.index')->with('success', 'Feeding report deleted successfully.');
